ml/requirements.txt
# Talisay Oil Yield Prediction - Python Dependencies

# Core ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep Learning
tensorflow>=2.15.0
# Alternative: torch>=2.0.0

# YOLO Object Detection (NEW)
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API Server
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0

# Image Augmentation
albumentations>=1.3.0

# Model Serialization
onnx>=1.14.0
onnxruntime>=1.16.0
